+++ src/main/java/sparta/ifour/movietalk/domain/reviews/controller/ReviewCudController.java
@@ -3,7 +3,6 @@ package sparta.ifour.movietalk.domain.reviews.controller;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 import sparta.ifour.movietalk.domain.reviews.dto.ReviewRequestDto;
-import sparta.ifour.movietalk.domain.reviews.dto.ReviewResponseDto;
 import sparta.ifour.movietalk.domain.reviews.service.ReviewService;
 
 /**
